🧬 Extract all the logic about loading templatable configuration into a new single module *_load*.
